Michael Obster <praenti@lug-in.de>: ddd: update to version 3.3.8 libmikmod: is not an active project at the moment -> URL has changed
openmotif:
2.2.2 has some bigger bugs and nedit makes problems with that library. 
The 2.1.30 is known as very stable, but also with bugs. But there is 
group at metrolink which have fixed the bugs in this version.
--> patch for downgrade openmotif to 2.1.30-4-MLI
--> patch to disable install wrapper, because this version doesn't 
produce shared files errors

nedit:
only tested with openmotif!!! Please test it also with lestiff!
update to version 5.4 (this version doesn't want openmotif 2.2.2, so 
please downgrade openmotif before)


git-svn-id: http://www.rocklinux.org/svn/rock-linux/trunk@1881 c5f82cb5-29bc-0310-9cd0-bff59a50e3bc