Added a "check for modified files" option to mine ("-y" for now ...).
I know that a:
md5sum --check var/adm/md5sums/$pkg

is sufficent for the check task - but a statically linked version for 
rescue situations are a nice-to-have (ehrm yes - I had to solve some stability
problems on my iBook in the last days ....). Mine should be the right place 
for a common package check interface.


git-svn-id: http://www.rocklinux.org/svn/rock-linux/trunk@2061 c5f82cb5-29bc-0310-9cd0-bff59a50e3bc